Support for DetectedIssue FHIR resources (#220)
* Adds QuestionnaireResponse processing API

Implements an API endpoint to receive QuestionnaireResponses, transform them into FHIR Observations, and post them to an external FHIR server.

This introduces a new controller, service, and builder classes to handle the conversion and submission process. It also fixes an issue with FHIR client connection when the base URL is a string.

* Adds rake tasks to fetch and submit FHIR bundles

Adds rake tasks for fetching QuestionnaireResponses from a FHIR server, transforming them into transaction bundles, and submitting these bundles back to the server.

This facilitates processing and submission of questionnaire data. Includes error handling and logging for failed submissions.

* Added support of collections in pfe_observation_builder.rb

* Added some initial code to support DetectedIssue FHIR resources.

* Refactor DetectedIssue and ModelHelper: Enhance provider name parsing, add codeable concept handling, and improve evidence and mitigation processing

* Enhance DetectedIssuesController and ResourceFetchHelper: Implement pagination for detected issues and add detected issue to patient-related resources; update locale for no detected issues message

* Add Detected Issues filtering and display: Implement detected issue count in ApplicationController, create detected_issue_filters_controller for filtering functionality, and add detected issues view with filter options and table display.

* Remove package-lock.json to clean up unused dependencies

* Update README.md: Revise supported FHIR resources and update Ruby version

---------

Co-authored-by: Vanessa Fotso <vfotso@mitre.org>